House_Price_prediction/requirements.txt
azureml-core
azureml-sdk
azureml-defaults
scikit-learn
pandas
gast
glob2
imutils
numpy
joblib
argparse